control_tower/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
docker==7.0.0
requests==2.31.0
urllib3<2
python-multipart==0.0.9
psutil==5.9.8
pandas==2.2.0
kafka-python==2.0.2
python-logging-loki==0.3.1
psycopg2-binary==2.9.9
